Update Main Nav to be mobile first
