+++ src/components/layout/hero/Hero.css
@@ -18,10 +18,6 @@
   max-width: 100%;
 }
 
-.hero-text h1 {
-  margin-left: 5px; /* keeps text from leaving box */
-}
-
 .hero-text {
   position: relative;
   display: flex;
@@ -36,7 +32,7 @@
 }
 
 .hero-text h1 {
-  margin: 10px 0 13px;
+  margin: 0 0 13px 10px;
 }
 
 .hero-code {
