+++ src/components/layout/navigation/MainNavigation.css
@@ -3,38 +3,63 @@
   height: 70px;
   background-color: #212121;
   display: flex;
-  justify-content: center;
+  justify-content: space-between;
   align-items: center;
-  position: relative;
-  top: 0px;
-  left: 0;
-  z-index: 99;
   padding: 0 25px;
 }
 
-.main-nav .logo {
+.nav-logo {
   width: 50px;
   margin-right: 10px;
 }
 
-.main-nav .logo a {
+.nav-logo a {
   display: flex;
   justify-content: center;
   align-items: center;
 }
 
-.main-nav ul {
+.nav-menu {
+  margin-top: 0; /* remove the margin from sibling divs*/
+}
+
+.nav-menu button {
+  font-size: 25px;
+  background: none;
+}
+
+.nav-links {
+  width: 100%;
+  max-width: 100%;
+  position: absolute;
+  top: 70px;
+  left: 0;
+  z-index: 5;
+  margin: 0;
+  background-color: #212121;
+  height: 100%;
+  max-height: 0;
+  overflow: hidden;
+  transition: all 0.4s cubic-bezier(0, 1, 0, 1);
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+}
+
+.nav-links.show {
+  max-height: 100vh;
+  transition: all 0.4s cubic-bezier(1, 0, 1, 0);
+}
+
+.nav-links ul {
   display: flex;
   width: 100%;
   max-width: 100%;
-  flex-direction: row;
-  justify-content: space-between;
+  flex-direction: column;
+  justify-content: space-evenly;
   align-items: center;
-  /* height: 70px;
-  background-color: #242424;
-  position: sticky;
-  top: 70px;
-  z-index: 0; */
+  height: 100%;
+  max-height: 75vh;
 }
 
 .main-nav ul li {
@@ -42,7 +67,7 @@
 }
 
 .main-nav ul a {
-  font-size: 18px;
+  font-size: 25px;
   text-decoration: none;
   color: rgba(255, 255, 255, 0.8);
 }
@@ -55,4 +80,32 @@
 .main-nav ul a:focus {
   color: rgba(200, 200, 200, 0.8);
   transition: color 0.2s ease-in;
+}
+
+@media screen and (min-width: 768px) {
+}
+
+@media screen and (min-width: 1024px) {
+  .nav-menu {
+    display: none;
+  }
+
+  .nav-links {
+    max-width: 600px;
+    position: relative;
+    max-height: 70px;
+    top: 0;
+    right: 0;
+  }
+
+  .nav-links ul {
+    max-height: 70px;
+    flex-direction: row;
+    justify-content: space-between;
+    position: relative;
+  }
+
+  .main-nav ul a {
+    font-size: 20px;
+  }
 }
