+++ src/components/layout/navigation/data/mainNavItems.js
@@ -1,0 +1,20 @@
+const mainNavItems = [
+  {
+    name: "About",
+    link: "#about",
+  },
+  {
+    name: "Experience",
+    link: "#experience",
+  },
+  {
+    name: "Projects",
+    link: "#projects",
+  },
+  {
+    name: "Contact",
+    link: "#contact",
+  },
+];
+
+export default mainNavItems;
